requirements.txt
streamlit==1.15.0
torch==1.10.0
torchvision==0.11.1
Pillow==8.4.0